// apps/web/src/api/generated/model/messageDtoStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Open WebUI Klon API
 * OpenAPI spec version: 0.0.0
 */

export type MessageDtoStatus = typeof MessageDtoStatus[keyof typeof MessageDtoStatus];


export const MessageDtoStatus = {
  complete: 'complete',
  aborted: 'aborted',
  error: 'error',
} as const;